providers/merge: alter auto_merges.status to int

+++ bot.sql
@@ -75,7 +75,7 @@ CREATE TABLE `auto_merges` (
   `repo` varchar(1023) COLLATE utf8_bin NOT NULL,
   `base_ref` varchar(1023) COLLATE utf8_bin NOT NULL,
   `started` tinyint(1) NOT NULL,
-  `status` tinyint(1) NOT NULL,
+  `status` int(11) NOT NULL,
   `created_at` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
   PRIMARY KEY (`id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_bin;
